Extract utilities to separate files
